refactor: extract figma helpers
